[process_builder] Zero out the bss even if it is contained within the last data page
Without this patch, the bss section is not zeroed out by the loader if
it fully lies within the last page of its PT_LOAD segment. As a result,
it will contain whatever is present in the ELF file after the last
non-bss byte.

This issue does not seem to affect any of Fuchsia's own ELF files that
are loaded by this function:
 - libc.so's bss section is (much) bigger than a single page
 - the vdso has no bss at all

However, this issue breaks Debian's ld-linux-x86-64.so.2 (when loaded by
starnix), whose loaded bss bytes will actually be initialized to the
contents of the DYNAMIC segment:
 Type           Offset             VirtAddr           PhysAddr
                FileSiz            MemSiz              Flags  Align
 ... [snip]
 LOAD           0x00000000000294c0 0x000000000002a4c0 0x000000000002a4c0
                0x0000000000001b40 0x0000000000001cb8  RW     0x1000
 DYNAMIC        0x0000000000029e78 0x000000000002ae78 0x000000000002ae78
                0x0000000000000170 0x0000000000000170  RW     0x8

Test: fx test process-builder-tests
      Verified that "gcc /tmp/hello_world.c" no longer pagefaults within "stardock run gcc"