[starnix] Use RLIMIT_STACK as stack size
To avoid the need to investigate this kind of page faults (i.e. increase
the hardcoded value and recompile iteratively until either the page
fault no longer occurs or it becomes apparent that it is not related to
the size of the stack). Since VMO pages are only committed on first
touch, this patch does not waste memory.

This is functionally equivalent to how real Linux behaves (apart from
the way the range is reported in maps and smaps files) as proven by
https://gist.github.com/fabio-d/8d33c79a88a879ff56bd0ab61b428464

Change-Id: I7bbb6fb2f44b88715db3ade1ff5a37ab7fa68573