[stardock] Support pulling non-official images too
Official images are those whose name starts with a "library/" prefix,
that is automatically added by the client (e.g. "busybox" is actually
"library/busybox").

This patch makes it possible to fetch images with a different prefix
too (by avoiding to prepend "library/" if the requested name already
has a slash in it).